Replace emoji with yes/no content
The emoji announce in screen readers as "X, graphic" or "white underscore check underscore mark, graphic" (TalkBack) or "crossmark" or "white heavy checkmark" (NVDA). And in Android Chrome, the emoji for the checkmark is white, which causes contrast issues. Using words will read faster and not have the contrast issue.